backend/requirements.txt
openai==0.28
python-dotenv==1.1.0
langchain
langchain-google-genai
imapclient
beautifulsoup4
chardet
google-generativeai==0.8.5
google-ai-generativelanguage==0.6.15
flask